cheat2
Added summoning and destroying of state objects. Handy stuff. As it
stands, you have to do any cheating when you first enter a node, after
that it can’t do it (which may or may not be worth having. As a debug
method it’s reasonably ok, as a cheat code less so.)